requirements.txt
numpy
torch
opencv-python
tqdm
Pillow
pathlib
argparse
rasterio
ultralytics
alphashape